refactor: Move trees to common `trees/` directory
