+++ parser/trees/tree_print.py
@@ -5,7 +5,7 @@ from io import StringIO
 from typing import IO, Sequence
 
 from .base_node import Leaf, AnyNode, Node
-from ..str_region import StrRegion
+from parser.str_region import StrRegion
 
 __all__ = [
     'TreePrinter', 'tree_print', 'tree_format', 'tprint', 'tformat'
