+++ test/common.py
@@ -5,8 +5,8 @@ from __future__ import annotations
 from enum import IntFlag, Enum
 from typing import Sequence, TypeVar
 
-from parser.cst.base_node import Leaf, AnyNode, Node
-from parser.cst.tree_print import tformat
+from parser.trees.base_node import Leaf, AnyNode, Node
+from parser.trees.tree_print import tformat
 from parser.cst.treegen import TreeGen, CstParseError
 from parser.error import BaseParseError
 from parser.lexer import Tokenizer
